+++ CavyGame/HomeViewController.swift
@@ -25,7 +25,7 @@ class HomeViewController: UIViewController, DMDynamicPageViewControllerDelegate,
     @IBOutlet weak var rightBarBtn: UIBarButtonItem!
     
     var viewController1 : RecommendationVC!
-    var viewController2 : RankListVC!
+    var viewController2 : RankingVC!
     var viewController3 : ClassificationTbVC!
     var viewController4 : PrefectureViewController!
     
@@ -185,7 +185,7 @@ class HomeViewController: UIViewController, DMDynamicPageViewControllerDelegate,
     func initPageController() {
         
         viewController1 = RecommendationVC()
-        viewController2 = RankListVC()
+        viewController2 = RankingVC()
         viewController3 = ClassificationTbVC()
         viewController4 = Common.getViewControllerWithIdentifier("Prefecture", identifier: "PrefectureView") as? PrefectureViewController
